Use pumpify to close remote socket (#5)
When the erp server closes the connection, the remote socket (connecting to the tunnel server) must be closed. The nodejs' buit-in socket.pipe doesn't close the other socket properly when one of the socket closes in the pipeline.
A popular library Pumpify fixes this issue.

Also, added a small testing server called http-echo-server for testing requests.